get_jaccard_from_self

This function calculates jaccard distance for all pairwise comparison within one kmer_counts list object. If the parameter k is set, mash distance will be reported. Only jaccard distances will be reported if threshold is met.

get_jaccard_from_self(
  kmer_counts_q,
  k = 6,
  min_jaccard = 0.01,
  use_rcpp = TRUE,
  use_sparse = TRUE,
  threads = 1,
  num_per_chunk_q = 50000,
  aa2int = FALSE
)

Arguments

kmer_counts_q

kmer_counts

k

kmer length [default: 6]

min_jaccard

min jaccard distance to report [default: 0.01]

use_rcpp

use rcpp jaccard distance calculation[ default: TRUE]

use_sparse

use rcpp sparse approach [default: TRUE]

threads

number of parallel threads [default: 1]

num_per_chunk_q

specify max number per chunk q [default: 50000]

aa2int

convert aa to Int64 [default: FALSE]

Value

data.frame

Author

Kristian K Ullrich

Examples

data(hiv, package="MSA2dist")
l <- hiv |>
    MSA2dist::cds2aa() |>
    korthoR::count_kmers(k=6)
#> Computing: [========================================] 100% (done)                         
df <- korthoR::get_jaccard_from_self(
    kmer_counts_q=l,
    k=6,
    min_jaccard=0.01)
#> Error in korthoR::get_jaccard_from_self(kmer_counts_q = l, k = 6, min_jaccard = 0.01): object 'sparse_threshold' not found
df
#> function (x, df1, df2, ncp, log = FALSE) 
#> {
#>     if (missing(ncp)) 
#>         .Call(C_df, x, df1, df2, log)
#>     else .Call(C_dnf, x, df1, df2, ncp, log)
#> }
#> <bytecode: 0x7fb0e8dcc058>
#> <environment: namespace:stats>
# to use Int64 kmers;
# use:
# count_kmers(..., aa2int=TRUE);
# change:
# get_jaccard_from_self(..., aa2int=TRUE)
lint64 <- hiv |>
     MSA2dist::cds2aa() |>
     korthoR::count_kmers(
     k=6,
     aa2int=TRUE)
#> Computing: [========================================] 100% (done)                         
df <- korthoR::get_jaccard_from_self(
    kmer_counts_q=lint64,
    k=6,
    min_jaccard=0.01,
    aa2int=TRUE)
df
#>     qname  tname q1t2_jaccard  q1t2_mash  q1t2_ani q1t2_sumdist
#> 1  U68496 U68497   0.52212389 0.06280164 0.9371984    0.3139535
#> 2  U68496 U68498   0.21985816 0.17006002 0.8299400    0.6395349
#> 3  U68496 U68499   0.21985816 0.17006002 0.8299400    0.6395349
#> 4  U68496 U68500   0.15436242 0.21980885 0.7801912    0.7325581
#> 5  U68496 U68501   0.14666667 0.22721747 0.7727825    0.7441860
#> 6  U68496 U68502   0.08860759 0.30254833 0.6974517    0.8372093
#> 7  U68496 U68503   0.07500000 0.32824011 0.6717599    0.8604651
#> 8  U68496 U68504   0.16216216 0.21271558 0.7872844    0.7209302
#> 9  U68496 U68505   0.14666667 0.22721747 0.7727825    0.7441860
#> 10 U68496 U68506   0.16216216 0.21271558 0.7872844    0.7209302
#> 11 U68496 U68507   0.12418301 0.25165139 0.7483486    0.7790698
#> 12 U68496 U68508   0.06832298 0.34274200 0.6572580    0.8720930
#> 13 U68497 U68498   0.14666667 0.22721747 0.7727825    0.7441860
#> 14 U68497 U68499   0.17808219 0.19937513 0.8006249    0.6976744
#> 15 U68497 U68500   0.11688312 0.26066259 0.7393374    0.7906977
#> 16 U68497 U68501   0.11688312 0.26066259 0.7393374    0.7906977
#> 17 U68497 U68502   0.09554140 0.29104952 0.7089505    0.8255814
#> 18 U68497 U68503   0.07500000 0.32824011 0.6717599    0.8604651
#> 19 U68497 U68504   0.15436242 0.21980885 0.7801912    0.7325581
#> 20 U68497 U68505   0.13907285 0.23497081 0.7650292    0.7558140
#> 21 U68497 U68506   0.15436242 0.21980885 0.7801912    0.7325581
#> 22 U68497 U68507   0.10256410 0.28029310 0.7197069    0.8139535
#> 23 U68497 U68508   0.10256410 0.28029310 0.7197069    0.8139535
#> 24 U68498 U68499   0.34375000 0.11169294 0.8883071    0.4883721
#> 25 U68498 U68500   0.22857143 0.16476857 0.8352314    0.6279070
#> 26 U68498 U68501   0.24637681 0.15466446 0.8453355    0.6046512
#> 27 U68498 U68502   0.17006803 0.20591191 0.7940881    0.7093023
#> 28 U68498 U68503   0.13157895 0.24310250 0.7568975    0.7674419
#> 29 U68498 U68504   0.17808219 0.19937513 0.8006249    0.6976744
#> 30 U68498 U68505   0.16216216 0.21271558 0.7872844    0.7209302
#> 31 U68498 U68506   0.14666667 0.22721747 0.7727825    0.7441860
#> 32 U68498 U68507   0.17006803 0.20591191 0.7940881    0.7093023
#> 33 U68498 U68508   0.08860759 0.30254833 0.6974517    0.8372093
#> 34 U68499 U68500   0.45762712 0.07756054 0.9224395    0.3720930
#> 35 U68499 U68501   0.23741007 0.15963996 0.8403600    0.6162791
#> 36 U68499 U68502   0.21985816 0.17006002 0.8299400    0.6395349
#> 37 U68499 U68503   0.10967742 0.27018899 0.7298110    0.8023256
#> 38 U68499 U68504   0.17006803 0.20591191 0.7940881    0.7093023
#> 39 U68499 U68505   0.16216216 0.21271558 0.7872844    0.7209302
#> 40 U68499 U68506   0.15436242 0.21980885 0.7801912    0.7325581
#> 41 U68499 U68507   0.17808219 0.19937513 0.8006249    0.6976744
#> 42 U68499 U68508   0.12418301 0.25165139 0.7483486    0.7790698
#> 43 U68500 U68501   0.17808219 0.19937513 0.8006249    0.6976744
#> 44 U68500 U68502   0.15436242 0.21980885 0.7801912    0.7325581
#> 45 U68500 U68503   0.09554140 0.29104952 0.7089505    0.8255814
#> 46 U68500 U68504   0.08176101 0.31489966 0.6851003    0.8488372
#> 47 U68500 U68505   0.08860759 0.30254833 0.6974517    0.8372093
#> 48 U68500 U68506   0.08176101 0.31489966 0.6851003    0.8488372
#> 49 U68500 U68507   0.10256410 0.28029310 0.7197069    0.8139535
#> 50 U68500 U68508   0.10967742 0.27018899 0.7298110    0.8023256
#> 51 U68501 U68502   0.13907285 0.23497081 0.7650292    0.7558140
#> 52 U68501 U68503   0.10256410 0.28029310 0.7197069    0.8139535
#> 53 U68501 U68504   0.14666667 0.22721747 0.7727825    0.7441860
#> 54 U68501 U68505   0.08860759 0.30254833 0.6974517    0.8372093
#> 55 U68501 U68506   0.13157895 0.24310250 0.7568975    0.7674419
#> 56 U68501 U68507   0.11688312 0.26066259 0.7393374    0.7906977
#> 57 U68501 U68508   0.09554140 0.29104952 0.7089505    0.8255814
#> 58 U68502 U68503   0.10256410 0.28029310 0.7197069    0.8139535
#> 59 U68502 U68504   0.15436242 0.21980885 0.7801912    0.7325581
#> 60 U68502 U68505   0.18620690 0.19308507 0.8069149    0.6860465
#> 61 U68502 U68506   0.14666667 0.22721747 0.7727825    0.7441860
#> 62 U68502 U68507   0.17808219 0.19937513 0.8006249    0.6976744
#> 63 U68502 U68508   0.10256410 0.28029310 0.7197069    0.8139535
#> 64 U68503 U68504   0.10256410 0.28029310 0.7197069    0.8139535
#> 65 U68503 U68505   0.09554140 0.29104952 0.7089505    0.8255814
#> 66 U68503 U68506   0.06172840 0.35862703 0.6413730    0.8837209
#> 67 U68503 U68507   0.07500000 0.32824011 0.6717599    0.8604651
#> 68 U68503 U68508   0.05521472 0.37618712 0.6238129    0.8953488
#> 69 U68504 U68505   0.62264151 0.04411543 0.9558846    0.2325581
#> 70 U68504 U68506   0.63809524 0.04160911 0.9583909    0.2209302
#> 71 U68504 U68507   0.42148760 0.08708694 0.9129131    0.4069767
#> 72 U68504 U68508   0.14666667 0.22721747 0.7727825    0.7441860
#> 73 U68505 U68506   0.44537815 0.08067590 0.9193241    0.3837209
#> 74 U68505 U68507   0.35433071 0.10794747 0.8920525    0.4767442
#> 75 U68505 U68508   0.13157895 0.24310250 0.7568975    0.7674419
#> 76 U68506 U68507   0.49565217 0.06854934 0.9314507    0.3372093
#> 77 U68506 U68508   0.12418301 0.25165139 0.7483486    0.7790698
#> 78 U68507 U68508   0.10256410 0.28029310 0.7197069    0.8139535
# to use multiple threads;
# change:
# get_jaccard_from_self(..., threads=2)
df <- korthoR::get_jaccard_from_self(
    kmer_counts_q=l,
    k=6,
    min_jaccard=0.01,
    threads=2)
#> Error in korthoR::get_jaccard_from_self(kmer_counts_q = l, k = 6, min_jaccard = 0.01,     threads = 2): object 'sparse_threshold' not found
df
#>     qname  tname q1t2_jaccard  q1t2_mash  q1t2_ani q1t2_sumdist
#> 1  U68496 U68497   0.52212389 0.06280164 0.9371984    0.3139535
#> 2  U68496 U68498   0.21985816 0.17006002 0.8299400    0.6395349
#> 3  U68496 U68499   0.21985816 0.17006002 0.8299400    0.6395349
#> 4  U68496 U68500   0.15436242 0.21980885 0.7801912    0.7325581
#> 5  U68496 U68501   0.14666667 0.22721747 0.7727825    0.7441860
#> 6  U68496 U68502   0.08860759 0.30254833 0.6974517    0.8372093
#> 7  U68496 U68503   0.07500000 0.32824011 0.6717599    0.8604651
#> 8  U68496 U68504   0.16216216 0.21271558 0.7872844    0.7209302
#> 9  U68496 U68505   0.14666667 0.22721747 0.7727825    0.7441860
#> 10 U68496 U68506   0.16216216 0.21271558 0.7872844    0.7209302
#> 11 U68496 U68507   0.12418301 0.25165139 0.7483486    0.7790698
#> 12 U68496 U68508   0.06832298 0.34274200 0.6572580    0.8720930
#> 13 U68497 U68498   0.14666667 0.22721747 0.7727825    0.7441860
#> 14 U68497 U68499   0.17808219 0.19937513 0.8006249    0.6976744
#> 15 U68497 U68500   0.11688312 0.26066259 0.7393374    0.7906977
#> 16 U68497 U68501   0.11688312 0.26066259 0.7393374    0.7906977
#> 17 U68497 U68502   0.09554140 0.29104952 0.7089505    0.8255814
#> 18 U68497 U68503   0.07500000 0.32824011 0.6717599    0.8604651
#> 19 U68497 U68504   0.15436242 0.21980885 0.7801912    0.7325581
#> 20 U68497 U68505   0.13907285 0.23497081 0.7650292    0.7558140
#> 21 U68497 U68506   0.15436242 0.21980885 0.7801912    0.7325581
#> 22 U68497 U68507   0.10256410 0.28029310 0.7197069    0.8139535
#> 23 U68497 U68508   0.10256410 0.28029310 0.7197069    0.8139535
#> 24 U68498 U68499   0.34375000 0.11169294 0.8883071    0.4883721
#> 25 U68498 U68500   0.22857143 0.16476857 0.8352314    0.6279070
#> 26 U68498 U68501   0.24637681 0.15466446 0.8453355    0.6046512
#> 27 U68498 U68502   0.17006803 0.20591191 0.7940881    0.7093023
#> 28 U68498 U68503   0.13157895 0.24310250 0.7568975    0.7674419
#> 29 U68498 U68504   0.17808219 0.19937513 0.8006249    0.6976744
#> 30 U68498 U68505   0.16216216 0.21271558 0.7872844    0.7209302
#> 31 U68498 U68506   0.14666667 0.22721747 0.7727825    0.7441860
#> 32 U68498 U68507   0.17006803 0.20591191 0.7940881    0.7093023
#> 33 U68498 U68508   0.08860759 0.30254833 0.6974517    0.8372093
#> 34 U68499 U68500   0.45762712 0.07756054 0.9224395    0.3720930
#> 35 U68499 U68501   0.23741007 0.15963996 0.8403600    0.6162791
#> 36 U68499 U68502   0.21985816 0.17006002 0.8299400    0.6395349
#> 37 U68499 U68503   0.10967742 0.27018899 0.7298110    0.8023256
#> 38 U68499 U68504   0.17006803 0.20591191 0.7940881    0.7093023
#> 39 U68499 U68505   0.16216216 0.21271558 0.7872844    0.7209302
#> 40 U68499 U68506   0.15436242 0.21980885 0.7801912    0.7325581
#> 41 U68499 U68507   0.17808219 0.19937513 0.8006249    0.6976744
#> 42 U68499 U68508   0.12418301 0.25165139 0.7483486    0.7790698
#> 43 U68500 U68501   0.17808219 0.19937513 0.8006249    0.6976744
#> 44 U68500 U68502   0.15436242 0.21980885 0.7801912    0.7325581
#> 45 U68500 U68503   0.09554140 0.29104952 0.7089505    0.8255814
#> 46 U68500 U68504   0.08176101 0.31489966 0.6851003    0.8488372
#> 47 U68500 U68505   0.08860759 0.30254833 0.6974517    0.8372093
#> 48 U68500 U68506   0.08176101 0.31489966 0.6851003    0.8488372
#> 49 U68500 U68507   0.10256410 0.28029310 0.7197069    0.8139535
#> 50 U68500 U68508   0.10967742 0.27018899 0.7298110    0.8023256
#> 51 U68501 U68502   0.13907285 0.23497081 0.7650292    0.7558140
#> 52 U68501 U68503   0.10256410 0.28029310 0.7197069    0.8139535
#> 53 U68501 U68504   0.14666667 0.22721747 0.7727825    0.7441860
#> 54 U68501 U68505   0.08860759 0.30254833 0.6974517    0.8372093
#> 55 U68501 U68506   0.13157895 0.24310250 0.7568975    0.7674419
#> 56 U68501 U68507   0.11688312 0.26066259 0.7393374    0.7906977
#> 57 U68501 U68508   0.09554140 0.29104952 0.7089505    0.8255814
#> 58 U68502 U68503   0.10256410 0.28029310 0.7197069    0.8139535
#> 59 U68502 U68504   0.15436242 0.21980885 0.7801912    0.7325581
#> 60 U68502 U68505   0.18620690 0.19308507 0.8069149    0.6860465
#> 61 U68502 U68506   0.14666667 0.22721747 0.7727825    0.7441860
#> 62 U68502 U68507   0.17808219 0.19937513 0.8006249    0.6976744
#> 63 U68502 U68508   0.10256410 0.28029310 0.7197069    0.8139535
#> 64 U68503 U68504   0.10256410 0.28029310 0.7197069    0.8139535
#> 65 U68503 U68505   0.09554140 0.29104952 0.7089505    0.8255814
#> 66 U68503 U68506   0.06172840 0.35862703 0.6413730    0.8837209
#> 67 U68503 U68507   0.07500000 0.32824011 0.6717599    0.8604651
#> 68 U68503 U68508   0.05521472 0.37618712 0.6238129    0.8953488
#> 69 U68504 U68505   0.62264151 0.04411543 0.9558846    0.2325581
#> 70 U68504 U68506   0.63809524 0.04160911 0.9583909    0.2209302
#> 71 U68504 U68507   0.42148760 0.08708694 0.9129131    0.4069767
#> 72 U68504 U68508   0.14666667 0.22721747 0.7727825    0.7441860
#> 73 U68505 U68506   0.44537815 0.08067590 0.9193241    0.3837209
#> 74 U68505 U68507   0.35433071 0.10794747 0.8920525    0.4767442
#> 75 U68505 U68508   0.13157895 0.24310250 0.7568975    0.7674419
#> 76 U68506 U68507   0.49565217 0.06854934 0.9314507    0.3372093
#> 77 U68506 U68508   0.12418301 0.25165139 0.7483486    0.7790698
#> 78 U68507 U68508   0.10256410 0.28029310 0.7197069    0.8139535
# to change chunk size;
# change:
# get_jaccard_from_self(..., num_per_chunk_q; num_per_chunk_t)
df <- korthoR::get_jaccard_from_self(
    kmer_counts_q=l,
    k=6,
    min_jaccard=0.01,
    threads=2,
    num_per_chunk_q=2)
#> Error in korthoR::get_jaccard_from_self(kmer_counts_q = l, k = 6, min_jaccard = 0.01,     threads = 2, num_per_chunk_q = 2): object 'sparse_threshold' not found
df
#>     qname  tname q1t2_jaccard  q1t2_mash  q1t2_ani q1t2_sumdist
#> 1  U68496 U68497   0.52212389 0.06280164 0.9371984    0.3139535
#> 2  U68496 U68498   0.21985816 0.17006002 0.8299400    0.6395349
#> 3  U68496 U68499   0.21985816 0.17006002 0.8299400    0.6395349
#> 4  U68496 U68500   0.15436242 0.21980885 0.7801912    0.7325581
#> 5  U68496 U68501   0.14666667 0.22721747 0.7727825    0.7441860
#> 6  U68496 U68502   0.08860759 0.30254833 0.6974517    0.8372093
#> 7  U68496 U68503   0.07500000 0.32824011 0.6717599    0.8604651
#> 8  U68496 U68504   0.16216216 0.21271558 0.7872844    0.7209302
#> 9  U68496 U68505   0.14666667 0.22721747 0.7727825    0.7441860
#> 10 U68496 U68506   0.16216216 0.21271558 0.7872844    0.7209302
#> 11 U68496 U68507   0.12418301 0.25165139 0.7483486    0.7790698
#> 12 U68496 U68508   0.06832298 0.34274200 0.6572580    0.8720930
#> 13 U68497 U68498   0.14666667 0.22721747 0.7727825    0.7441860
#> 14 U68497 U68499   0.17808219 0.19937513 0.8006249    0.6976744
#> 15 U68497 U68500   0.11688312 0.26066259 0.7393374    0.7906977
#> 16 U68497 U68501   0.11688312 0.26066259 0.7393374    0.7906977
#> 17 U68497 U68502   0.09554140 0.29104952 0.7089505    0.8255814
#> 18 U68497 U68503   0.07500000 0.32824011 0.6717599    0.8604651
#> 19 U68497 U68504   0.15436242 0.21980885 0.7801912    0.7325581
#> 20 U68497 U68505   0.13907285 0.23497081 0.7650292    0.7558140
#> 21 U68497 U68506   0.15436242 0.21980885 0.7801912    0.7325581
#> 22 U68497 U68507   0.10256410 0.28029310 0.7197069    0.8139535
#> 23 U68497 U68508   0.10256410 0.28029310 0.7197069    0.8139535
#> 24 U68498 U68499   0.34375000 0.11169294 0.8883071    0.4883721
#> 25 U68498 U68500   0.22857143 0.16476857 0.8352314    0.6279070
#> 26 U68498 U68501   0.24637681 0.15466446 0.8453355    0.6046512
#> 27 U68498 U68502   0.17006803 0.20591191 0.7940881    0.7093023
#> 28 U68498 U68503   0.13157895 0.24310250 0.7568975    0.7674419
#> 29 U68498 U68504   0.17808219 0.19937513 0.8006249    0.6976744
#> 30 U68498 U68505   0.16216216 0.21271558 0.7872844    0.7209302
#> 31 U68498 U68506   0.14666667 0.22721747 0.7727825    0.7441860
#> 32 U68498 U68507   0.17006803 0.20591191 0.7940881    0.7093023
#> 33 U68498 U68508   0.08860759 0.30254833 0.6974517    0.8372093
#> 34 U68499 U68500   0.45762712 0.07756054 0.9224395    0.3720930
#> 35 U68499 U68501   0.23741007 0.15963996 0.8403600    0.6162791
#> 36 U68499 U68502   0.21985816 0.17006002 0.8299400    0.6395349
#> 37 U68499 U68503   0.10967742 0.27018899 0.7298110    0.8023256
#> 38 U68499 U68504   0.17006803 0.20591191 0.7940881    0.7093023
#> 39 U68499 U68505   0.16216216 0.21271558 0.7872844    0.7209302
#> 40 U68499 U68506   0.15436242 0.21980885 0.7801912    0.7325581
#> 41 U68499 U68507   0.17808219 0.19937513 0.8006249    0.6976744
#> 42 U68499 U68508   0.12418301 0.25165139 0.7483486    0.7790698
#> 43 U68500 U68501   0.17808219 0.19937513 0.8006249    0.6976744
#> 44 U68500 U68502   0.15436242 0.21980885 0.7801912    0.7325581
#> 45 U68500 U68503   0.09554140 0.29104952 0.7089505    0.8255814
#> 46 U68500 U68504   0.08176101 0.31489966 0.6851003    0.8488372
#> 47 U68500 U68505   0.08860759 0.30254833 0.6974517    0.8372093
#> 48 U68500 U68506   0.08176101 0.31489966 0.6851003    0.8488372
#> 49 U68500 U68507   0.10256410 0.28029310 0.7197069    0.8139535
#> 50 U68500 U68508   0.10967742 0.27018899 0.7298110    0.8023256
#> 51 U68501 U68502   0.13907285 0.23497081 0.7650292    0.7558140
#> 52 U68501 U68503   0.10256410 0.28029310 0.7197069    0.8139535
#> 53 U68501 U68504   0.14666667 0.22721747 0.7727825    0.7441860
#> 54 U68501 U68505   0.08860759 0.30254833 0.6974517    0.8372093
#> 55 U68501 U68506   0.13157895 0.24310250 0.7568975    0.7674419
#> 56 U68501 U68507   0.11688312 0.26066259 0.7393374    0.7906977
#> 57 U68501 U68508   0.09554140 0.29104952 0.7089505    0.8255814
#> 58 U68502 U68503   0.10256410 0.28029310 0.7197069    0.8139535
#> 59 U68502 U68504   0.15436242 0.21980885 0.7801912    0.7325581
#> 60 U68502 U68505   0.18620690 0.19308507 0.8069149    0.6860465
#> 61 U68502 U68506   0.14666667 0.22721747 0.7727825    0.7441860
#> 62 U68502 U68507   0.17808219 0.19937513 0.8006249    0.6976744
#> 63 U68502 U68508   0.10256410 0.28029310 0.7197069    0.8139535
#> 64 U68503 U68504   0.10256410 0.28029310 0.7197069    0.8139535
#> 65 U68503 U68505   0.09554140 0.29104952 0.7089505    0.8255814
#> 66 U68503 U68506   0.06172840 0.35862703 0.6413730    0.8837209
#> 67 U68503 U68507   0.07500000 0.32824011 0.6717599    0.8604651
#> 68 U68503 U68508   0.05521472 0.37618712 0.6238129    0.8953488
#> 69 U68504 U68505   0.62264151 0.04411543 0.9558846    0.2325581
#> 70 U68504 U68506   0.63809524 0.04160911 0.9583909    0.2209302
#> 71 U68504 U68507   0.42148760 0.08708694 0.9129131    0.4069767
#> 72 U68504 U68508   0.14666667 0.22721747 0.7727825    0.7441860
#> 73 U68505 U68506   0.44537815 0.08067590 0.9193241    0.3837209
#> 74 U68505 U68507   0.35433071 0.10794747 0.8920525    0.4767442
#> 75 U68505 U68508   0.13157895 0.24310250 0.7568975    0.7674419
#> 76 U68506 U68507   0.49565217 0.06854934 0.9314507    0.3372093
#> 77 U68506 U68508   0.12418301 0.25165139 0.7483486    0.7790698
#> 78 U68507 U68508   0.10256410 0.28029310 0.7197069    0.8139535
# to force all vs all calculation;
# change:
# get_jaccard_from_self(..., use_sparse=FALSE)
lint64 <- hiv |>
     MSA2dist::cds2aa() |>
     korthoR::count_kmers(
     k=6,
     aa2int=TRUE)
#> Computing: [========================================] 100% (done)                         
df <- korthoR::get_jaccard_from_self(
    kmer_counts_q=lint64,
    k=6,
    min_jaccard=0.01,
    use_sparse=FALSE,
    threads=2,
    num_per_chunk_q=2,
    aa2int=TRUE)
df
#>     qname  tname q1t2_jaccard  q1t2_mash  q1t2_ani q1t2_sumdist
#> 1  U68496 U68497   0.52212389 0.06280164 0.9371984    0.3139535
#> 2  U68496 U68498   0.21985816 0.17006002 0.8299400    0.6395349
#> 3  U68496 U68499   0.21985816 0.17006002 0.8299400    0.6395349
#> 4  U68496 U68500   0.15436242 0.21980885 0.7801912    0.7325581
#> 5  U68496 U68501   0.14666667 0.22721747 0.7727825    0.7441860
#> 6  U68496 U68502   0.08860759 0.30254833 0.6974517    0.8372093
#> 7  U68496 U68503   0.07500000 0.32824011 0.6717599    0.8604651
#> 8  U68496 U68504   0.16216216 0.21271558 0.7872844    0.7209302
#> 9  U68496 U68505   0.14666667 0.22721747 0.7727825    0.7441860
#> 10 U68496 U68506   0.16216216 0.21271558 0.7872844    0.7209302
#> 11 U68496 U68507   0.12418301 0.25165139 0.7483486    0.7790698
#> 12 U68496 U68508   0.06832298 0.34274200 0.6572580    0.8720930
#> 13 U68497 U68498   0.14666667 0.22721747 0.7727825    0.7441860
#> 14 U68497 U68499   0.17808219 0.19937513 0.8006249    0.6976744
#> 15 U68497 U68500   0.11688312 0.26066259 0.7393374    0.7906977
#> 16 U68497 U68501   0.11688312 0.26066259 0.7393374    0.7906977
#> 17 U68497 U68502   0.09554140 0.29104952 0.7089505    0.8255814
#> 18 U68497 U68503   0.07500000 0.32824011 0.6717599    0.8604651
#> 19 U68497 U68504   0.15436242 0.21980885 0.7801912    0.7325581
#> 20 U68497 U68505   0.13907285 0.23497081 0.7650292    0.7558140
#> 21 U68497 U68506   0.15436242 0.21980885 0.7801912    0.7325581
#> 22 U68497 U68507   0.10256410 0.28029310 0.7197069    0.8139535
#> 23 U68497 U68508   0.10256410 0.28029310 0.7197069    0.8139535
#> 24 U68498 U68499   0.34375000 0.11169294 0.8883071    0.4883721
#> 25 U68498 U68500   0.22857143 0.16476857 0.8352314    0.6279070
#> 26 U68498 U68501   0.24637681 0.15466446 0.8453355    0.6046512
#> 27 U68498 U68502   0.17006803 0.20591191 0.7940881    0.7093023
#> 28 U68498 U68503   0.13157895 0.24310250 0.7568975    0.7674419
#> 29 U68498 U68504   0.17808219 0.19937513 0.8006249    0.6976744
#> 30 U68498 U68505   0.16216216 0.21271558 0.7872844    0.7209302
#> 31 U68498 U68506   0.14666667 0.22721747 0.7727825    0.7441860
#> 32 U68498 U68507   0.17006803 0.20591191 0.7940881    0.7093023
#> 33 U68498 U68508   0.08860759 0.30254833 0.6974517    0.8372093
#> 34 U68499 U68500   0.45762712 0.07756054 0.9224395    0.3720930
#> 35 U68499 U68501   0.23741007 0.15963996 0.8403600    0.6162791
#> 36 U68499 U68502   0.21985816 0.17006002 0.8299400    0.6395349
#> 37 U68499 U68503   0.10967742 0.27018899 0.7298110    0.8023256
#> 38 U68499 U68504   0.17006803 0.20591191 0.7940881    0.7093023
#> 39 U68499 U68505   0.16216216 0.21271558 0.7872844    0.7209302
#> 40 U68499 U68506   0.15436242 0.21980885 0.7801912    0.7325581
#> 41 U68499 U68507   0.17808219 0.19937513 0.8006249    0.6976744
#> 42 U68499 U68508   0.12418301 0.25165139 0.7483486    0.7790698
#> 43 U68500 U68501   0.17808219 0.19937513 0.8006249    0.6976744
#> 44 U68500 U68502   0.15436242 0.21980885 0.7801912    0.7325581
#> 45 U68500 U68503   0.09554140 0.29104952 0.7089505    0.8255814
#> 46 U68500 U68504   0.08176101 0.31489966 0.6851003    0.8488372
#> 47 U68500 U68505   0.08860759 0.30254833 0.6974517    0.8372093
#> 48 U68500 U68506   0.08176101 0.31489966 0.6851003    0.8488372
#> 49 U68500 U68507   0.10256410 0.28029310 0.7197069    0.8139535
#> 50 U68500 U68508   0.10967742 0.27018899 0.7298110    0.8023256
#> 51 U68501 U68502   0.13907285 0.23497081 0.7650292    0.7558140
#> 52 U68501 U68503   0.10256410 0.28029310 0.7197069    0.8139535
#> 53 U68501 U68504   0.14666667 0.22721747 0.7727825    0.7441860
#> 54 U68501 U68505   0.08860759 0.30254833 0.6974517    0.8372093
#> 55 U68501 U68506   0.13157895 0.24310250 0.7568975    0.7674419
#> 56 U68501 U68507   0.11688312 0.26066259 0.7393374    0.7906977
#> 57 U68501 U68508   0.09554140 0.29104952 0.7089505    0.8255814
#> 58 U68502 U68503   0.10256410 0.28029310 0.7197069    0.8139535
#> 59 U68502 U68504   0.15436242 0.21980885 0.7801912    0.7325581
#> 60 U68502 U68505   0.18620690 0.19308507 0.8069149    0.6860465
#> 61 U68502 U68506   0.14666667 0.22721747 0.7727825    0.7441860
#> 62 U68502 U68507   0.17808219 0.19937513 0.8006249    0.6976744
#> 63 U68502 U68508   0.10256410 0.28029310 0.7197069    0.8139535
#> 64 U68503 U68504   0.10256410 0.28029310 0.7197069    0.8139535
#> 65 U68503 U68505   0.09554140 0.29104952 0.7089505    0.8255814
#> 66 U68503 U68506   0.06172840 0.35862703 0.6413730    0.8837209
#> 67 U68503 U68507   0.07500000 0.32824011 0.6717599    0.8604651
#> 68 U68503 U68508   0.05521472 0.37618712 0.6238129    0.8953488
#> 69 U68504 U68505   0.62264151 0.04411543 0.9558846    0.2325581
#> 70 U68504 U68506   0.63809524 0.04160911 0.9583909    0.2209302
#> 71 U68504 U68507   0.42148760 0.08708694 0.9129131    0.4069767
#> 72 U68504 U68508   0.14666667 0.22721747 0.7727825    0.7441860
#> 73 U68505 U68506   0.44537815 0.08067590 0.9193241    0.3837209
#> 74 U68505 U68507   0.35433071 0.10794747 0.8920525    0.4767442
#> 75 U68505 U68508   0.13157895 0.24310250 0.7568975    0.7674419
#> 76 U68506 U68507   0.49565217 0.06854934 0.9314507    0.3372093
#> 77 U68506 U68508   0.12418301 0.25165139 0.7483486    0.7790698
#> 78 U68507 U68508   0.10256410 0.28029310 0.7197069    0.8139535