rcpp_jaccard_sparse_Int64

returns jaccard

Arguments

kmer_counts_q_Int64

list [mandatory]

kmer_counts_t_Int64

list [mandatory]

k

kmer length [default: 6]

min_jaccard

min jaccard distance to report [default: 0.01]

sparse_threshold

use kmer subset to evaluate change search strategy [default: 0.1]

sparse_n

number of kmer subset to get sparse threshold [default: 20]

ncores

number of cores [default: 1]

debug

print to console [default: FALSE]

Value

list

Author

Kristian K Ullrich

Examples

## load example sequence data
data("hiv", package="MSA2dist")
lint64 <- hiv |>
    MSA2dist::cds2aa() |>
    korthoR::count_kmers(
    k=6,
    aa2int=TRUE)
#> Computing: [========================================] 100% (done)                         
d <- korthoR::rcpp_jaccard_sparse_Int64(
    kmer_counts_q=lint64,
    kmer_counts_t=lint64,
    k=6,
    min_jaccard=0.01,
    sparse_threshold=0.1,
    sparse_n=100,
    ncores=1,
    debug=FALSE)
d
#>      qname  tname q1t2_jaccard  q1t2_mash  q1t2_ani q1t2_sumdist
#> 1   U68496 U68496   1.00000000 0.00000000 1.0000000    0.0000000
#> 2   U68496 U68497   0.52212389 0.06280164 0.9371984    0.3139535
#> 3   U68496 U68498   0.21985816 0.17006002 0.8299400    0.6395349
#> 4   U68496 U68499   0.21985816 0.17006002 0.8299400    0.6395349
#> 5   U68496 U68500   0.15436242 0.21980885 0.7801912    0.7325581
#> 6   U68496 U68501   0.14666667 0.22721747 0.7727825    0.7441860
#> 7   U68496 U68502   0.08860759 0.30254833 0.6974517    0.8372093
#> 8   U68496 U68503   0.07500000 0.32824011 0.6717599    0.8604651
#> 9   U68496 U68504   0.16216216 0.21271558 0.7872844    0.7209302
#> 10  U68496 U68505   0.14666667 0.22721747 0.7727825    0.7441860
#> 11  U68496 U68506   0.16216216 0.21271558 0.7872844    0.7209302
#> 12  U68496 U68507   0.12418301 0.25165139 0.7483486    0.7790698
#> 13  U68496 U68508   0.06832298 0.34274200 0.6572580    0.8720930
#> 14  U68497 U68496   0.52212389 0.06280164 0.9371984    0.3139535
#> 15  U68497 U68497   1.00000000 0.00000000 1.0000000    0.0000000
#> 16  U68497 U68498   0.14666667 0.22721747 0.7727825    0.7441860
#> 17  U68497 U68499   0.17808219 0.19937513 0.8006249    0.6976744
#> 18  U68497 U68500   0.11688312 0.26066259 0.7393374    0.7906977
#> 19  U68497 U68501   0.11688312 0.26066259 0.7393374    0.7906977
#> 20  U68497 U68502   0.09554140 0.29104952 0.7089505    0.8255814
#> 21  U68497 U68503   0.07500000 0.32824011 0.6717599    0.8604651
#> 22  U68497 U68504   0.15436242 0.21980885 0.7801912    0.7325581
#> 23  U68497 U68505   0.13907285 0.23497081 0.7650292    0.7558140
#> 24  U68497 U68506   0.15436242 0.21980885 0.7801912    0.7325581
#> 25  U68497 U68507   0.10256410 0.28029310 0.7197069    0.8139535
#> 26  U68497 U68508   0.10256410 0.28029310 0.7197069    0.8139535
#> 27  U68498 U68496   0.21985816 0.17006002 0.8299400    0.6395349
#> 28  U68498 U68497   0.14666667 0.22721747 0.7727825    0.7441860
#> 29  U68498 U68498   1.00000000 0.00000000 1.0000000    0.0000000
#> 30  U68498 U68499   0.34375000 0.11169294 0.8883071    0.4883721
#> 31  U68498 U68500   0.22857143 0.16476857 0.8352314    0.6279070
#> 32  U68498 U68501   0.24637681 0.15466446 0.8453355    0.6046512
#> 33  U68498 U68502   0.17006803 0.20591191 0.7940881    0.7093023
#> 34  U68498 U68503   0.13157895 0.24310250 0.7568975    0.7674419
#> 35  U68498 U68504   0.17808219 0.19937513 0.8006249    0.6976744
#> 36  U68498 U68505   0.16216216 0.21271558 0.7872844    0.7209302
#> 37  U68498 U68506   0.14666667 0.22721747 0.7727825    0.7441860
#> 38  U68498 U68507   0.17006803 0.20591191 0.7940881    0.7093023
#> 39  U68498 U68508   0.08860759 0.30254833 0.6974517    0.8372093
#> 40  U68499 U68496   0.21985816 0.17006002 0.8299400    0.6395349
#> 41  U68499 U68497   0.17808219 0.19937513 0.8006249    0.6976744
#> 42  U68499 U68498   0.34375000 0.11169294 0.8883071    0.4883721
#> 43  U68499 U68499   1.00000000 0.00000000 1.0000000    0.0000000
#> 44  U68499 U68500   0.45762712 0.07756054 0.9224395    0.3720930
#> 45  U68499 U68501   0.23741007 0.15963996 0.8403600    0.6162791
#> 46  U68499 U68502   0.21985816 0.17006002 0.8299400    0.6395349
#> 47  U68499 U68503   0.10967742 0.27018899 0.7298110    0.8023256
#> 48  U68499 U68504   0.17006803 0.20591191 0.7940881    0.7093023
#> 49  U68499 U68505   0.16216216 0.21271558 0.7872844    0.7209302
#> 50  U68499 U68506   0.15436242 0.21980885 0.7801912    0.7325581
#> 51  U68499 U68507   0.17808219 0.19937513 0.8006249    0.6976744
#> 52  U68499 U68508   0.12418301 0.25165139 0.7483486    0.7790698
#> 53  U68500 U68496   0.15436242 0.21980885 0.7801912    0.7325581
#> 54  U68500 U68497   0.11688312 0.26066259 0.7393374    0.7906977
#> 55  U68500 U68498   0.22857143 0.16476857 0.8352314    0.6279070
#> 56  U68500 U68499   0.45762712 0.07756054 0.9224395    0.3720930
#> 57  U68500 U68500   1.00000000 0.00000000 1.0000000    0.0000000
#> 58  U68500 U68501   0.17808219 0.19937513 0.8006249    0.6976744
#> 59  U68500 U68502   0.15436242 0.21980885 0.7801912    0.7325581
#> 60  U68500 U68503   0.09554140 0.29104952 0.7089505    0.8255814
#> 61  U68500 U68504   0.08176101 0.31489966 0.6851003    0.8488372
#> 62  U68500 U68505   0.08860759 0.30254833 0.6974517    0.8372093
#> 63  U68500 U68506   0.08176101 0.31489966 0.6851003    0.8488372
#> 64  U68500 U68507   0.10256410 0.28029310 0.7197069    0.8139535
#> 65  U68500 U68508   0.10967742 0.27018899 0.7298110    0.8023256
#> 66  U68501 U68496   0.14666667 0.22721747 0.7727825    0.7441860
#> 67  U68501 U68497   0.11688312 0.26066259 0.7393374    0.7906977
#> 68  U68501 U68498   0.24637681 0.15466446 0.8453355    0.6046512
#> 69  U68501 U68499   0.23741007 0.15963996 0.8403600    0.6162791
#> 70  U68501 U68500   0.17808219 0.19937513 0.8006249    0.6976744
#> 71  U68501 U68501   1.00000000 0.00000000 1.0000000    0.0000000
#> 72  U68501 U68502   0.13907285 0.23497081 0.7650292    0.7558140
#> 73  U68501 U68503   0.10256410 0.28029310 0.7197069    0.8139535
#> 74  U68501 U68504   0.14666667 0.22721747 0.7727825    0.7441860
#> 75  U68501 U68505   0.08860759 0.30254833 0.6974517    0.8372093
#> 76  U68501 U68506   0.13157895 0.24310250 0.7568975    0.7674419
#> 77  U68501 U68507   0.11688312 0.26066259 0.7393374    0.7906977
#> 78  U68501 U68508   0.09554140 0.29104952 0.7089505    0.8255814
#> 79  U68502 U68496   0.08860759 0.30254833 0.6974517    0.8372093
#> 80  U68502 U68497   0.09554140 0.29104952 0.7089505    0.8255814
#> 81  U68502 U68498   0.17006803 0.20591191 0.7940881    0.7093023
#> 82  U68502 U68499   0.21985816 0.17006002 0.8299400    0.6395349
#> 83  U68502 U68500   0.15436242 0.21980885 0.7801912    0.7325581
#> 84  U68502 U68501   0.13907285 0.23497081 0.7650292    0.7558140
#> 85  U68502 U68502   1.00000000 0.00000000 1.0000000    0.0000000
#> 86  U68502 U68503   0.10256410 0.28029310 0.7197069    0.8139535
#> 87  U68502 U68504   0.15436242 0.21980885 0.7801912    0.7325581
#> 88  U68502 U68505   0.18620690 0.19308507 0.8069149    0.6860465
#> 89  U68502 U68506   0.14666667 0.22721747 0.7727825    0.7441860
#> 90  U68502 U68507   0.17808219 0.19937513 0.8006249    0.6976744
#> 91  U68502 U68508   0.10256410 0.28029310 0.7197069    0.8139535
#> 92  U68503 U68496   0.07500000 0.32824011 0.6717599    0.8604651
#> 93  U68503 U68497   0.07500000 0.32824011 0.6717599    0.8604651
#> 94  U68503 U68498   0.13157895 0.24310250 0.7568975    0.7674419
#> 95  U68503 U68499   0.10967742 0.27018899 0.7298110    0.8023256
#> 96  U68503 U68500   0.09554140 0.29104952 0.7089505    0.8255814
#> 97  U68503 U68501   0.10256410 0.28029310 0.7197069    0.8139535
#> 98  U68503 U68502   0.10256410 0.28029310 0.7197069    0.8139535
#> 99  U68503 U68503   1.00000000 0.00000000 1.0000000    0.0000000
#> 100 U68503 U68504   0.10256410 0.28029310 0.7197069    0.8139535
#> 101 U68503 U68505   0.09554140 0.29104952 0.7089505    0.8255814
#> 102 U68503 U68506   0.06172840 0.35862703 0.6413730    0.8837209
#> 103 U68503 U68507   0.07500000 0.32824011 0.6717599    0.8604651
#> 104 U68503 U68508   0.05521472 0.37618712 0.6238129    0.8953488
#> 105 U68504 U68496   0.16216216 0.21271558 0.7872844    0.7209302
#> 106 U68504 U68497   0.15436242 0.21980885 0.7801912    0.7325581
#> 107 U68504 U68498   0.17808219 0.19937513 0.8006249    0.6976744
#> 108 U68504 U68499   0.17006803 0.20591191 0.7940881    0.7093023
#> 109 U68504 U68500   0.08176101 0.31489966 0.6851003    0.8488372
#> 110 U68504 U68501   0.14666667 0.22721747 0.7727825    0.7441860
#> 111 U68504 U68502   0.15436242 0.21980885 0.7801912    0.7325581
#> 112 U68504 U68503   0.10256410 0.28029310 0.7197069    0.8139535
#> 113 U68504 U68504   1.00000000 0.00000000 1.0000000    0.0000000
#> 114 U68504 U68505   0.62264151 0.04411543 0.9558846    0.2325581
#> 115 U68504 U68506   0.63809524 0.04160911 0.9583909    0.2209302
#> 116 U68504 U68507   0.42148760 0.08708694 0.9129131    0.4069767
#> 117 U68504 U68508   0.14666667 0.22721747 0.7727825    0.7441860
#> 118 U68505 U68496   0.14666667 0.22721747 0.7727825    0.7441860
#> 119 U68505 U68497   0.13907285 0.23497081 0.7650292    0.7558140
#> 120 U68505 U68498   0.16216216 0.21271558 0.7872844    0.7209302
#> 121 U68505 U68499   0.16216216 0.21271558 0.7872844    0.7209302
#> 122 U68505 U68500   0.08860759 0.30254833 0.6974517    0.8372093
#> 123 U68505 U68501   0.08860759 0.30254833 0.6974517    0.8372093
#> 124 U68505 U68502   0.18620690 0.19308507 0.8069149    0.6860465
#> 125 U68505 U68503   0.09554140 0.29104952 0.7089505    0.8255814
#> 126 U68505 U68504   0.62264151 0.04411543 0.9558846    0.2325581
#> 127 U68505 U68505   1.00000000 0.00000000 1.0000000    0.0000000
#> 128 U68505 U68506   0.44537815 0.08067590 0.9193241    0.3837209
#> 129 U68505 U68507   0.35433071 0.10794747 0.8920525    0.4767442
#> 130 U68505 U68508   0.13157895 0.24310250 0.7568975    0.7674419
#> 131 U68506 U68496   0.16216216 0.21271558 0.7872844    0.7209302
#> 132 U68506 U68497   0.15436242 0.21980885 0.7801912    0.7325581
#> 133 U68506 U68498   0.14666667 0.22721747 0.7727825    0.7441860
#> 134 U68506 U68499   0.15436242 0.21980885 0.7801912    0.7325581
#> 135 U68506 U68500   0.08176101 0.31489966 0.6851003    0.8488372
#> 136 U68506 U68501   0.13157895 0.24310250 0.7568975    0.7674419
#> 137 U68506 U68502   0.14666667 0.22721747 0.7727825    0.7441860
#> 138 U68506 U68503   0.06172840 0.35862703 0.6413730    0.8837209
#> 139 U68506 U68504   0.63809524 0.04160911 0.9583909    0.2209302
#> 140 U68506 U68505   0.44537815 0.08067590 0.9193241    0.3837209
#> 141 U68506 U68506   1.00000000 0.00000000 1.0000000    0.0000000
#> 142 U68506 U68507   0.49565217 0.06854934 0.9314507    0.3372093
#> 143 U68506 U68508   0.12418301 0.25165139 0.7483486    0.7790698
#> 144 U68507 U68496   0.12418301 0.25165139 0.7483486    0.7790698
#> 145 U68507 U68497   0.10256410 0.28029310 0.7197069    0.8139535
#> 146 U68507 U68498   0.17006803 0.20591191 0.7940881    0.7093023
#> 147 U68507 U68499   0.17808219 0.19937513 0.8006249    0.6976744
#> 148 U68507 U68500   0.10256410 0.28029310 0.7197069    0.8139535
#> 149 U68507 U68501   0.11688312 0.26066259 0.7393374    0.7906977
#> 150 U68507 U68502   0.17808219 0.19937513 0.8006249    0.6976744
#> 151 U68507 U68503   0.07500000 0.32824011 0.6717599    0.8604651
#> 152 U68507 U68504   0.42148760 0.08708694 0.9129131    0.4069767
#> 153 U68507 U68505   0.35433071 0.10794747 0.8920525    0.4767442
#> 154 U68507 U68506   0.49565217 0.06854934 0.9314507    0.3372093
#> 155 U68507 U68507   1.00000000 0.00000000 1.0000000    0.0000000
#> 156 U68507 U68508   0.10256410 0.28029310 0.7197069    0.8139535
#> 157 U68508 U68496   0.06832298 0.34274200 0.6572580    0.8720930
#> 158 U68508 U68497   0.10256410 0.28029310 0.7197069    0.8139535
#> 159 U68508 U68498   0.08860759 0.30254833 0.6974517    0.8372093
#> 160 U68508 U68499   0.12418301 0.25165139 0.7483486    0.7790698
#> 161 U68508 U68500   0.10967742 0.27018899 0.7298110    0.8023256
#> 162 U68508 U68501   0.09554140 0.29104952 0.7089505    0.8255814
#> 163 U68508 U68502   0.10256410 0.28029310 0.7197069    0.8139535
#> 164 U68508 U68503   0.05521472 0.37618712 0.6238129    0.8953488
#> 165 U68508 U68504   0.14666667 0.22721747 0.7727825    0.7441860
#> 166 U68508 U68505   0.13157895 0.24310250 0.7568975    0.7674419
#> 167 U68508 U68506   0.12418301 0.25165139 0.7483486    0.7790698
#> 168 U68508 U68507   0.10256410 0.28029310 0.7197069    0.8139535
#> 169 U68508 U68508   1.00000000 0.00000000 1.0000000    0.0000000